Moved switching session back to finally block. Some cleanup
