+++ entity-store/src/main/kotlin/jetbrains/exodus/entitystore/orientdb/iterate/property/OPropertyEqualIterable.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package jetbrains.exodus.entitystore.iterate.property
+package jetbrains.exodus.entitystore.orientdb.iterate.property
 
 import jetbrains.exodus.entitystore.*
 import jetbrains.exodus.entitystore.orientdb.iterate.OQueryEntityIterableBase
