+++ query/src/main/java/jetbrains/exodus/query/GetLinks.java
@@ -18,7 +18,7 @@ package jetbrains.exodus.query;
 
 import jetbrains.exodus.entitystore.Entity;
 import jetbrains.exodus.entitystore.EntityId;
-import jetbrains.exodus.entitystore.iterate.property.OPropertyEqualIterable;
+import jetbrains.exodus.entitystore.orientdb.iterate.property.OPropertyEqualIterable;
 import jetbrains.exodus.entitystore.orientdb.iterate.OQueryEntityIterableBase;
 import jetbrains.exodus.query.metadata.ModelMetaData;
